+++ src-tauri/Cargo.toml
@@ -15,10 +15,11 @@ tauri = { version = "=2.0.0-alpha.10", features = [] }
 tauri-plugin-fs = { git = "https://github.com/tauri-apps/plugins-workspace", branch = "v2" }
 tauri-plugin-window = { git = "https://github.com/tauri-apps/plugins-workspace", branch = "v2" }
 tauri-plugin-os = { git = "https://github.com/tauri-apps/plugins-workspace", branch = "v2" }
+diesel = { version = "2.1.0", features = ["sqlite"] }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 rodio = "0.17.1"
-lofty = "0.13.0"
+lofty = "0.14.0"
 walkdir = "2.3.3"
 
 [features]
