+++ src-tauri/src/database/schema.rs
@@ -1,0 +1,8 @@
+// @generated automatically by Diesel CLI.
+
+diesel::table! {
+    library (id) {
+        id -> Integer,
+        path -> Text,
+    }
+}
