+++ src-tauri/src/lib.rs
@@ -5,11 +5,9 @@ mod mobile;
 pub use mobile::*;
 
 mod audio;
+mod database;
 
-use crate::audio::{PlaybackState, tracks::fetch_tracks};
-use audio::{play_file, queue_file, toggle_pause};
-use tauri::{path::BaseDirectory, App, Manager};
-use tauri_plugin_fs::FsExt;
+use tauri::App;
 
 //noinspection RsWrongGenericArgumentsNumber
 pub type SetupHook = Box<dyn FnOnce(&mut App) -> Result<(), Box<dyn std::error::Error>> + Send>;
@@ -40,33 +38,18 @@ impl AppBuilder {
             .plugin(tauri_plugin_fs::init())
             .plugin(tauri_plugin_window::init())
             .plugin(tauri_plugin_os::init())
-            .invoke_handler(tauri::generate_handler![play_file, queue_file, toggle_pause])
+            // Make commands invokable from frontend
+            .invoke_handler(tauri::generate_handler![])
             .setup(move |app| {
                 if let Some(setup) = setup {
                     (setup)(app)?;
                 }
 
-                // Allow fs to access system audio directory
-                let audio_path = app.path().resolve("Horizon", BaseDirectory::Audio)?;
-                if std::fs::read_dir(&audio_path).err().is_some() {
-                    std::fs::create_dir(&audio_path).expect("Could not create audio directory")
-                }
-                app.fs_scope().allow_directory(&audio_path, true)?;
+                // Set up the database
+                database::setup(app.handle());
 
-                // Setup playback state
-                PlaybackState::setup(app.handle());
-
-                // Listen for client mounted event
-                let app_handle = app.handle();
-                app.listen_global("mounted", move |event| {
-                    println!("Front has mounted. Fetching tracks and status...");
-
-                    // Send tracks to client
-                    let tracks = fetch_tracks(audio_path.clone()).expect("Failed to fetch tracks");
-                    app_handle
-                        .emit_all("tracks", tracks)
-                        .expect("Could not emit tracks");
-                });
+                // Set up the backend
+                audio::setup(app.handle());
 
                 Ok(())
             })
